+++ kernels/builders/heuristic_binning.h
@@ -60,9 +60,9 @@ namespace embree
         {
           const vint4 i = floori((vfloat4(p)-ofs)*scale);
 #if 1
-          assert(i[0] >=0 && i[0] < num); 
-          assert(i[1] >=0 && i[1] < num);
-          assert(i[2] >=0 && i[2] < num);
+          assert(i[0] >= 0 && (size_t)i[0] < num); 
+		  assert(i[1] >= 0 && (size_t)i[1] < num);
+		  assert(i[2] >= 0 && (size_t)i[2] < num);
           return Vec3ia(i);
 #else
           return Vec3ia(clamp(i,vint4(0),vint4(num-1)));
@@ -364,7 +364,7 @@ namespace embree
 	
 	size_t leftCount = 0;
 	BBox3fa leftBounds = empty;
-	for (size_t i=0; i<split.pos; i++) {
+	for (size_t i=0; i<(size_t)split.pos; i++) {
 	  leftCount += counts[i][split.dim];
 	  leftBounds.extend(bounds[i][split.dim]);
 	}
